Ensure ordering for particles in tests
as per
https://pytest-xdist.readthedocs.io/en/latest/known-limitations.html#workarounds,
it turns out you simply need to wrap your argument lists in `sorted`
(I'm not entirely sure why a list is not good enough; but maybe sorting
them on a per-worker basis does... something. I'm not sure what).

I also moved `particle`, which just returns a particle from the particle
zoo, to a subpackage-shared fixture.
