+++ plasmapy/particles/tests/test_parsing.py
@@ -331,16 +331,17 @@ def test_parse_InvalidParticleErrors(arg, kwargs):
         )
 
 
-@pytest.mark.parametrize("arg", particle_zoo.everything - {"p+"})
-def test_parse_InvalidElementErrors(arg):
+def test_parse_InvalidElementErrors(particle):
     r"""Tests that _parse_and_check_atomic_input raises an
     InvalidElementError when the input corresponds to a valid
     particle but not a valid element, isotope, or ion."""
+    if particle == Particle("p+"):
+        return  # TODO verify this case should still be a skip
     with pytest.raises(InvalidElementError):
-        parse_and_check_atomic_input(arg)
+        parse_and_check_atomic_input(particle)
         pytest.fail(
             "An InvalidElementError was expected to be raised by "
-            f"{call_string(parse_and_check_atomic_input, arg)}, "
+            f"{call_string(parse_and_check_atomic_input, particle)}, "
             f"but no exception was raised."
         )
